OneDrive/Escritorio/EGINMOTEG/backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.1
django-environ>=0.11
django-simple-history>=3.5
django-storages[s3]>=1.14
django-celery-beat>=2.6
psycopg2-binary>=2.9
Pillow>=10.2
celery[redis]>=5.3
redis>=5.0
gunicorn>=21.2
drf-spectacular>=0.27
scikit-learn>=1.4
numpy>=1.26
joblib>=1.3
openai>=1.12
